Refactor module tests.test_client
Drastically simplify the tests for wrapper methods such as
`nailgun.client.delete`. Make `ContentTypeIsJsonTestCase` a bit more thorough
and easy to read. Merge the contents of `ArgTestCase` in to `ClientTestCase` and
drop all references to the `_call_requests_*` methods. Drop all references to
DDT and drop it as a requirement.