Drop `nailgun.client._call_requests_*` methods
The `_call_requests_*` methods were originally created with the intent of making
tests easier to write. The idea is that those methods were small, simple pieces
of code that sat at the edge of this application and the requsts library.
However, the methods are not actually needed. The standard library tools for
mocking make it easy to directly alter how requests methods behave. Drop them.