Release version 0.3.0
This release is the first in a series of tightly related releases. This one adds
the `nailgun.entity_fields` module, and later releases will add the
`nailgun.entity_mixins` and `nailgun.entities` modules. Changes in this release:

* Port all *Field classes from Robottelo to module `nailgun.entity_fields`.
* Improve documentation. Most notably, stop documeting special class members
  such as `__init__` and `__weakref__`.

Beware that methods `OneToOneField.gen_value` and `OneToManyField.gen_value`
will, under certain circumstances, throw an exception. For example:

    OneToOneField('MyEntity').gen_value()

The problem here is that `gen_value` will try to find a class named "MyEntity"
in module `nailgun.entities`, but that module does not exist yet. Until the
entity classes are ported from Robottelo to Nailgun, this issue can be worked
around using one of the following methods:

* Pass a class to the `OneToOneField` or `OneToManyField` constructor, thus
  obviating the need to go searching for a class in the first place.
* Pass the `module` parameter to the `OneToOneField` or `OneToManyField`
  constructor, thus specifying where class "MyEntity" can be found.
* Set a new default location for finding classes by setting
  `nailgun.entity_fields.ENTITIES_MODULE`.