Fix numerous broken links in the documentation
When Sphinx processes a `:param:` or `:rtype:` directive, it will try to link to
the referenced object. For example, consider the following docstring snippet:

    :param dict kwargs: The keyword args supplied to :func:`request` or one of
        the convenience functions like it.
    :returns: ``True`` or ``False``
    :rtype: bool

In this case, Sphinx will try to find and link to the objects "dict" and "bool".
Unfortunately, those objects do not exist in the NailGun code base. This issue
can be masked by using the `nitpick_ignore` option ion the Sphinx configuration
file, but Sphinx will still produce malformed documentation.

Fix this by not linking to objects that cannot be referenced. Prevent this issue
from silently creeping back in to the code base by dropping the `nitpick_ignore`
option from Sphinx's configuration file.