Drop misleading logging statements
Module `nailgun.client` logs out information about each request that is issues
and each response that is received. Here is an example of some logging
statements:

    nailgun.client: DEBUG: Making HTTP GET request to
    https://example.com/api/v2/permissions with options {'verify': False,
    'data': '{"per_page": 10000, "name": "upload_abrt_reports"}', 'auth': (…),
    'headers': {'content-type': 'application/json'}} and no data.

    nailgun.client: DEBUG: Equivalent curl command: curl -X GET --user …
    --insecure
    data=%7B%22per_page%22%3A+10000%2C+%22name%22%3A+%22upload_abrt_reports%22%7D&headers=%7B%27content-type%27%3A+%27application%2Fjson%27%7D
    https://example.com/api/v2/permissions

Unfortunately, the "equivalent curl command" is not, in fact, equivalent. The
format of the data is pretty well messed up and the request headers are placed
in completely the wrong location.

In addition, these logging statements make it harder to read through logs
after-the-fact. NailGun produces a verbose amount of logging, and these
statements are arguably excessive enough to be counter-productive.

Drop curl-related logging statements.