Fix OneTo{One,Many}Field.gen_value
The instance methods `OneToOneField.gen_value` and `OneToManyField.gen_value`
return objects. For example, if a `OneToOneField` references a `User` class,
then calling `gen_value` on it will return a `User` object. This is problematic,
because it is sometimes desireable or necessary to customize the process by
which new entity objects are created. In addition, new entity objects require
that a `ServerConfig` object be passed in, but cannot provide an appropriate
server configuration! As an alternative, client code has to fall back on
`_get_class`.

Fix this by making `OneToOneField.gen_value` and `OneToManyField.gen_value`
return classes, not objects. Client code can then pass in whatever values are
appropriate for the object being instantiated, and client code can pass in an
appropate server configuration object too. This obviates the need for client
code to call `_get_class`.

Fix #25.

Example test results against a downstream system with this code:

    $ nosetests tests/foreman/api/test_multiple_paths.py
    ...S...S.....S.............................................................S...S.......................................................................EE...........................................................SSSS.SSSSSS..S.SSSSSSS...SSSSSSSSSSSSSSSSSSSSSSSSSSSSSS
    Ran 267 tests in 439.863s

    FAILED (SKIP=53, errors=2)

The two test failures are due to a known issue with API versioning differences.