Release version 0.13.0
This release focuses on improving usability. It is now possible to easily
inspect the contents of `Entity` and `ServerConfig` objects, additional
documentation on `ServerConfig` objects is available, spurious warnings about
insecure connections are no longer displayed, only iterable values can be
assigned to `OneToManyField`s and the logging system properly captures and
emits messages.

Changes in this release:

* Add the "role" field to the "User" entity
* Check values assigned to `OneToManyField`s (#55)
* Customize how Entity and ServerConfig objects are stringified (#53)
* Make NailGun compatible with Requests 2.7.0
* Initialize the logging system (#61)
* Provide examples of using ServerConfig objects (#44)
* Give generated entities a ServerConfig object (#63)
* Silence spurious insecure connection warnings
* Make `Host.create` work
* Add unit tests for version-related entity logic