Drop the inner `Meta` class from entities
Fix #30:

> The inner Meta classes are arguably a bad idea. A class should contain a
> bundle of data and logic that operates on that data. However, the Entity
> classes contain only data. By this reasoning, it is better to drop the Meta
> classes and add a regular class attribute to hold the same data.

A more serious issue is that the use of an inner class brings shared state to
all entity classes. One way in which this is problematic is that is prevents
safe customization of an entity's API path. One class that does this is
`ContentViewPuppetModule`. It has API paths in the following form:

    /content_views/:content_view_id/content_view_puppet_modules
    /content_views/:content_view_id/content_view_puppet_modules/:id

When a `ContentViewPuppetModule` is instantiated, it sets its API path to
something like the following:

    /content_views/42/content_view_puppet_modules

And because an entity's API path is stored as a value on an inner class, this
change is reflected in *all* `ContentViewPuppetModule` objects. That's a bug.

Test results against a downstream system:

    $ nosetests tests/foreman/api/test_multiple_paths.py
    ...S...S......S...............................................................S...S..........................................................................................................................................SSS..S.S.SSSSSS.S..SS.SSSSSSSSSSSSSSSSSSSSSSSSSSSS
    ----------------------------------------------------------------------
    Ran 271 tests in 473.785s

    OK (SKIP=47)

These test results do require a change be made to Robottelo. Robottelo has one
reference to the inner `Meta` class in module `test_multiple_entities`.