Unit test overridden read() methods
One thing that the default `read` method does is to pass the current entity
object's server configuration to the entity object being created. Both parent
and child entity objects should talk to the same server with the same
credentials and so on. Unfortunately, several of the `read` method overrides in
module `nailgun.entities` fail to do that. Add a unit test for these methods to
ensure that they do so.

`ContentViewFilterRule.read` calls `keys` on a dict, then iterates through those
keys and modifies the underlying dict. This works on Python 2 where `keys`
returns an entirely new object, but it fails in Python 3 where `keys` returns a
view object:

    $ make test
    …
    Traceback (most recent call last):
    File "/home/ichimonji10/code/nailgun/tests/test_entities.py", line 435, in test_sync_plan
        entity.read()
    File "/home/ichimonji10/code/nailgun/nailgun/entities.py", line 842, in read
        for entity_field in entity._fields.keys():
    RuntimeError: dictionary changed size during iteration

Add unit tests for several of the overridden `read` methods and fix these
issues.