Update documentation
Currently, all of the modules in the `nailgun` package are documented in a
single reST file. Similarly, all of the modules in the `tests` package are
documented in a single reST file. This technically works, but the compiled
documentation contains some enormously long pages, and the navigation trees are
not intuitive. Fix this by breaking documentation down so there is one reST file
per module. As a result, the compiled documentation contains smaller pages and
more intuitive navigation trees.

Rework the actual documentation contents in several other locations. Provide a
brief explanation of how the methods in each mixin cooperate to do work, provide
a better introduction to the documentation at the root of the API documentation,
and fix a link in the "Examples" documentation.