Drop unnecessary entity classes
Drop the `HostClasses` and `HostGroupClasses` entity classes. The first class
exposes the following calls:

* GET    /api/hosts/:host_id/puppetclass_ids
* POST   /api/hosts/:host_id/puppetclass_ids
* DELETE /api/hosts/:host_id/puppetclass_ids/:id

The second class exposes the following calls:

* GET    /api/hostgroups/:hostgroup_id/puppetclass_ids
* POST   /api/hostgroups/:hostgroup_id/puppetclass_ids
* DELETE /api/hostgroups/:hostgroup_id/puppetclass_ids/:id

These calls appear quite useful, but in practice, their presence is confusing an
unnecessary. Consider `GET /api/hosts/:host_id/puppetclass_ids`. It is one of
three highly similar paths:

1. GET /api/puppetclasses
2. GET /api/hosts/:host_id/puppetclasses
3. GET /api/hosts/:host_id/puppetclass_ids

The first call returns all puppet classes by default. You can filter down
exactly how many puppet classes are returned by passing in a JSON payload such
as `{'host_id': 42}`. The second call returns hashes of information about puppet
classes that belong to a host. And the third call returns IDs of puppet classes
that belong to a host.

As you can see, the first API call is the most powerful, the second API call is
less powerful and the third API call is least powerful. As a result, one can
just use the first API call and ignore the other two, and no functionality is
lost.

Similar comments apply to all of the other paths. You can update the set of
puppet classes by making an HTTP POST call to
`/api/hosts/:host_id/puppetclass_ids`. However, it is preferable to make an HTTP
PUT call to `/api/hosts/:id` with a JSON payload of `{'host':
{'puppet_class_ids': [1, 2, 3]}}`. This is preferable not only because you're
using the most general and powerful mechanism available, but also because a PUT
is more appropriate than  POST.

All the above is a nice design discussion, but how will NailGun provide clients
with the ability to do things like get puppet classes belonging to a host or
update the puppet classes belonging to a host? These two actions can be made
possible by writing the `EntitySearchMixin` and `EntityUpdateMixin` as described
in #11 and #12. Statements like the following should be made available:

    # Update the puppet classes belonging to a host.
    host = entities.Host(id=1).read()
    host.puppet_classes.extend([PuppetClass(id=2), PuppetClass(id=3)])
    host.update(fields=['puppet_classes'])

    # Get puppet classes belonging to a host.
    hosts = entities.Host(puppet_classes=[2, 3]).search()