Fix AbstractComputeResource inheritance
The entity class `AbstractComputeResource` defines a set of fields that all
inheriting classes use. For example, all compute resources have a name and
provider. Child classes should be able to extend that set of fields. For
example, `LibvirtComputeResource` adds the `display_type` and
`set_console_password` fields.

Currently, that inheritance is broken. The following fails:

    from nailgun.entities import LibvirtComputeResource
    LibvirtComputeResource(display_type='VNC')

Similarly, any field which is defined on a child class cannot be assigned to
directly. Fix this.