Updated missing dependencies in Host.create_missing()
Some entities in Host.create_missing() were created without required
connections through organization and location.
API tests which are using create_missing() are passing without this fix, as
API doesn't check for requirements to be satisfied (so API allows to
associate entities from different locs and orgs, which is not good).
But UI tests are failing as corresponding entities aren't present in
drop-downs and can't be selected.
The fix is required for SatelliteQE/robottelo#748.