Update HostGroup entity
Add the `content_view` and `lifecycle_environment` fields to the `HostGroup`
entity. Both fields are covered by the API documentation, but neither field is
returned by an HTTP GET request to `/api/v2/hostgroups/:id`. In response, file
two bugs and add a workaround to the `read` method.

Add `EntityUpdateMixin` to the `HostGroup` entity. Add an `update` method that
works around two more bugs.

Update the test suite to fix tests that break as a result of these changes.