Make code more concise
As a result of #110 being fixed, it is now possible to take a few sections of
code and make them more concise.

The `HostGroup` entity appears to be in good order:

    >>> from nailgun import entities
    >>> host_group = entities.HostGroup(parent=1).create(True)
    >>> host_group.id
    50
    >>> host_group.parent.id
    u'1'
    >>> entities.HostGroup(id=50).read().parent.id
    u'1'
    >>> print(entities.HostGroup(id=1).read().parent)
    None

Robottelo is still happy with `HostGroup`:

    $ nosetests tests/foreman/api/test_multiple_paths.py -m HostGroup
    .........SS
    ----------------------------------------------------------------------
    Ran 11 tests in 10.644s

    OK (SKIP=2)

I am unable to completely verify that the `ContentViewPuppetModule` entity works
well, given that Robottelo does not currently use it in any capacity. However,
the change made there is extremely similar to the change made to `HostGroup.