Only require unittest 2 for Python 3.3 and below
Adjust the `requirements-dev.txt` file so that unittest2 is only installed if
Python 3.3 or below is being used. Adjust test modules to only import unittest2
if Python 3.3 or below is being used.

unittest2 is used on Python 3.3 and below so that the `subTest` context manager
can be used in unit tests. There is no need to use unittest2 on Python 3.4 and
above, as this feature has been added to the standard library.

Bonus: refactor out some unnecessary version-checking code.