Release version 0.20.0, classify as "Alpha"
There are very few commits in this release, but the few commits provide an
enormous amount of new functionality. This release is being made so that users
can start getting a handle on the new functionality as soon as possible.

Mark NailGun as being "Alpha" quality. This is in line with the milestone goals
listed on GitHub:

    https://github.com/SatelliteQE/nailgun/milestones/Make%20%22alpha%22%20Quality

Breaking changes in this release:

* Drop method `Permission.search`. Make class `EntitySearchMixin` a parent of
  class `Permission`. The mixin provides a method by the same name.

Changes in this release:

* Implement, test and document a search mixin. (#11)
* Add the search mixin to several entities:

  * Environment
  * GPGKey
  * LifecycleEnvironment
  * Location
  * Organization
  * Permission
  * Ping
  * Product
  * Role
  * User