Normalize entity helper methods
Fix #41:

> Several methods in module `nailgun.entities` do not map nicely to an
> underlying path. In addition, several methods do not have appropriate
> arguments. The methods exposed by the entity classes should be reviewed to
> make sure that they are appropriately named and placed on the most appropriate
> class.

Though many actions can be performed with the mixins, some do not map nicely to
one of the CRUD verbs. For example, publishing a content view or uploading
content to a repository are actions that cannot be described by the CRUD verbs.

NailGun must be able to carry out actions such as publishing or uploading. But
how? NailGun has thus far addressed these needs by simply adding methods on an
ad-hoc basis. This works, but it comes with costs:

* The helper methods must be learned one-by-one, as there is no guarantee of
  consistency among them.
* Some helper methods will require ongoing maintenance. For example, the
  `AbstractDockerContainer.power` method only accepts the strings "start,"
  "stop" and "status," and it will need to be updated when Satellite changes.
  NailGun already commits to creating entity classes for each resource and
  describing each available field — describing the logic of each helper method
  raises the cost of maintenance even higher.

Address these issues. Give each helper method a highly similar signature. Do not
attempt to interpret the messages sent from clients to the server, but instead
let the server validate the contents of each message.

The previous two commits in this chain of commits already make some head-way on
this front. This commit makes the following changes:

* Add the `Subscription` entity.
* Change the signature of `ActivationKey.content_override`.
* Change the signature of `Repository.upload_content`.
* Drop `ContentView.add_puppet_module`. (Use `ContentViewPuppetModule.create`
  instead.)
* Drop `ContentView.set_repository_ids`. (Use `ContentView.update` instead.)
* Drop `Organization.delete_manifest`. (Use `Subscription.delete_manifest`
  instead.)
* Drop `Organization.refresh_manifest`. (Use `Subscription.refresh_manifest`
  instead.)
* Drop `Organization.subscriptions`. (Use `Subscription.search` instead.)
* Drop `Organization.sync_plan`. (Use `SyncPlan.create` instead.)
* Drop `Organization.upload_manifest`. (Use `Subscription.upload` instead.)
* Drop `RHCIDeployment.add_hypervisors`. (Use RHCIDeployment.update` instead.)
* Drop `RepositorySet.list_all`. (Use `RepositorySet.search` instead.)
* Update unit tests.

Some of these methods may be re-added in some form or another. For example, it
would be useful to add a helper method `Organization.subscriptions_upload` that
maps to the `/katello/api/v2/organizations/:id/subscriptions/upload` path.
However, at this point in time, the focus is on making NailGun's API stable.
Developers should be able to write code against NailGun and be confident that it
will not break. These changes add stability, and future changes may build upon
them.