Add test case for helper methods
There are two kinds of methods on entities:

* Methods inherited from a mixin. For example, `ActivationKey.create` is
  inherited from the `EntityCreateMixin`.
* Methods defined on that entity, such as `ActivationKey.add_subscriptions`.

The latter set of methods was recently overhauled to make them more uniform. The
methods are now named after the path they call, they are placed on the correct
entities, and their signatures are very similar. However, there are still
differences. Some methods provide default values for their arguments; some
require a second argument; some require only one argument but of a different
kind than others; and so on. This is bad for several reasons:

* It is hard to learn the various methods. A user has to look at the signature
  of each individual method and figure out exactly what arguments it requires.
  This significantly raises the conceptual weight of the methods.
* It is hard to design the various methods. A developer has to spend time
  examining the API docs for the method call to figure out what information the
  API requires, the developer has to reference a variety of other helper methods
  to see which implementation should be copied, and so on.
* Tests for the various methods are not quite the same. This means that there's
  a lot of copy-cat code in the unit test suite.

Design a method signature that addresses these issues. It is an approach that
was not discussed in SatelliteQE/nailgun#41. Here are some examples of method
calls before the change:

    # ActivationKey.add_subscriptions
    activation_key.add_subscriptions({
        'quantity': 1,
        'subscription_id': subs.id,
    })

    # Subscription.upload
    with open(manifests.clone(), 'rb') as manifest:
        sub.upload({'organization_id': org.id}, manifest)

And after the change:

    # ActivationKey.add_subscriptions
    activation_key.add_subscriptions(data={
        'quantity': 1,
        'subscription_id': subs.id,
    })

    # Subscription.upload
    with open(manifests.clone(), 'rb') as handle:
        sub.upload(data={'organization_id': org.id}, files={'content': handle})

The updated approach is more verbose, but it's utterly uniform. All methods have
a signature like so:

    def method_name(self, synchronous=True, **kwargs):

Add a unit test for this new approach.