Hack around Pylint complaint
Pylint is configured to ignore duplicate code in module imports. See the recipe
for `make lint` for proof. However it is complaining about duplicate code in the
conditional import sections of code. Hack around this issue by re-ordering
imports.

An alternative approach is to add the `# pylint:disable=duplicate-code`
directive to the module. However, there is no clean way to apply this directive
to just the import section.