Change BaseServerConfig.__repr__
The `repr()` method provides an evaluatable string representation of an object.
The following should hold true:

    obj == exec(repr(obj))

Unfortunately, `BaseServerConfig.__repr__` fails to live up to that standard
whenever a version is passed in:

    >>> from nailgun.config import BaseServerConfig
    >>> BaseServerConfig('some url', version='1.2.3')
    nailgun.config.BaseServerConfig(url='some url', version=<Version('1.2.3')>)

Fix this issue. It is already true that when a `BaseServerConfig` object is
instantiated, the `version` argument is convered from a string to a `Version`
object. Make `BaseServerConfig.__repr__` convert the `version` attribute to a
string:

    >>> from nailgun.config import BaseServerConfig
    >>> cfg = BaseServerConfig('some url', version='1.2.3')
    >>> type(cfg.version)
    <class 'packaging.version.Version'>
    >>> cfg
    nailgun.config.BaseServerConfig(url='some url', version='1.2.3')