Make class ServerConfig set a default version
Fix #163.

When a `ServerConfig` object is created, only the 'url' argument is required.
All other arguments are optional. This is awkward when performing version
comparisons:

    >>> from nailgun.config import ServerConfig
    >>> hasattr(ServerConfig('some url'), 'version')
    False

This leads to idioms like the following in module `nailgun.entities`:

    if _get_version(self._server_config) < Version('6.1'):
        # do something

What does `_get_version` do? It returns the `version` attribute if present, or
assumes that the Satellite is the latest development version and returns version
'1!0'.

This sort of work-around is silly, because it forces users of the `version`
attribute to make assumptions about what version the Satellite server is, and
then scatters those assumptions throughout the code-base. It would be much
better to make this assumption once, in the `ServerConfig` object.

Before change:

    >>> from nailgun.config import ServerConfig
    >>> hasattr(ServerConfig('some url'), 'version')
    False

After change:

    >>> from nailgun.config import ServerConfig
    >>> cfg = ServerConfig('some url')
    >>> hasattr(cfg, 'version')
    True
    >>> cfg.version
    <Version('1!0')>