Remove codecov
This continues a trend we've done with other projects.
Codecov has some use in other types of projects, but really isn't a good
fit for Nailgun since it is functionally exercised by Robottelo.

Without getting too deep into philosophy, CodeCov is a misleading metric
for quality and minor changes are largely ignored in mature projects.